# public/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/scores": {
            "get": {
                "tags": [
                    "Score"
                ],
                "summary": "Get list of all scores with associated user and game",
                "operationId": "8ccac5260c190d3c079fec64df059a23",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Score"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Score"
                ],
                "summary": "Create a new score",
                "operationId": "048581179957f5d1e62afee858d20466",
                "requestBody": {
                    "description": "Provide score details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "userid",
                                    "gameid",
                                    "score"
                                ],
                                "properties": {
                                    "userid": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "gameid": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "score": {
                                        "type": "integer",
                                        "example": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Score created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Score"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/scores/{id}": {
            "get": {
                "tags": [
                    "Score"
                ],
                "summary": "Get a specific score with associated user and game",
                "operationId": "08f0fd0edcc2083731fdaa56c9c9eed1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Score ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Score found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Score"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Score not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Score"
                ],
                "summary": "Update an existing score",
                "operationId": "5729011f04d35479b947fa035c90484a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Score ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Score data to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Score"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Score updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Score"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Score not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Score"
                ],
                "summary": "Delete a score",
                "operationId": "d1190a7ece8cafa70ae84e7c0a1b0c06",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Score ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Score deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Score deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Score not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Game": {
                "title": "Game",
                "description": "Game model",
                "properties": {
                    "id": {
                        "description": "ID of the game",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nom": {
                        "description": "Name of the game",
                        "type": "string",
                        "example": "Chess"
                    }
                },
                "type": "object"
            },
...
            "Score": {
                "title": "Score",
                "description": "Score model",
                "properties": {
                    "id": {
                        "description": "ID of the score",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "userid": {
                        "description": "ID of the user",
                        "type": "integer",
                        "example": 1
                    },
                    "gameid": {
                        "description": "ID of the game",
                        "type": "integer",
                        "example": 1
                    },
                    "score": {
                        "description": "User's score in the game",
                        "type": "integer",
                        "example": 500
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "game": {
                        "$ref": "#/components/schemas/Game"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "ID of the user",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "nom": {
                        "description": "Username",
                        "type": "string",
                        "example": "johndoe"
                    },
                    "password": {
                        "description": "Password",
                        "type": "string",
                        "example": "pass1234"
                    }
                },
                "type": "object"
            }
        }
    }
}